# main/docs/swagger.yaml
basePath: /
definitions:
...
  dto.PageResponse:
    properties:
      page_no:
        description: 当前页码
        type: integer
      page_size:
        description: 每页大小
        type: integer
      total:
        description: 总数
        type: integer
    type: object
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  req.CallReq:
    properties:
      call_type:
        description: 呼叫类型1-服务员；2-结账
        enum:
        - 1
        - 2
        type: integer
      desk_uuid:
        description: 桌台ID
        type: integer
    required:
    - call_type
    - desk_uuid
    type: object
...
  req.PrinterLogReq:
    properties:
      uuid:
        description: 打印日志Uuid
        type: integer
    required:
    - uuid
    type: object
  req.ProcessedCallReq:
    properties:
      uuid:
        type: integer
    required:
    - uuid
    type: object
...
  resp.AbnormalPrintItem:
    properties:
      create_time:
        description: 创建时间
        type: integer
      desk_no:
        description: 桌台编号
        type: string
      printer_name:
        description: 打印时间
        type: string
      printer_uuid:
        description: 打印机Uuid
        type: integer
      reason:
        description: 异常原因
        type: string
      sale_bill_uuid:
        description: 销售账单uuid
        type: integer
      uuid:
        description: 打印日志Uuid
        type: integer
    type: object
  resp.AbnormalPrintList:
    properties:
      list:
        description: 异常打印列表
        items:
          $ref: '#/definitions/resp.AbnormalPrintItem'
        type: array
      meta:
        $ref: '#/definitions/dto.PageResponse'
    type: object
...
  resp.ReprintResp:
    properties:
      data:
        type: string
      print_method:
        description: 打印方式 1文本打印, 2图片打印
        type: integer
      print_times:
        type: integer
      printer_config:
        type: string
      printer_log_uuid:
        type: integer
      printer_name:
        type: string
      printer_time:
        type: integer
      printer_type:
        type: string
      printer_uuid:
        type: integer
    type: object
...
  resp.UnprocessedCallItem:
    properties:
      call_type:
        description: 呼叫类型:呼叫类型(1服务员,2结账)
        type: integer
      desk_no:
        description: 桌台编号
        type: string
      desk_uuid:
        description: 桌台Uuid
        type: integer
      is_send:
        description: 是否已发送：1-是；0-否
        type: integer
      uuid:
        description: 呼叫Uuid
        type: integer
    type: object
  resp.UnprocessedCallList:
    properties:
      list:
        description: 未处理呼叫列表
        items:
          $ref: '#/definitions/resp.UnprocessedCallItem'
        type: array
      meta:
        $ref: '#/definitions/dto.PageResponse'
    type: object
  resp.UnprocessedResp:
    properties:
      abnormal_print_count:
        description: 异常打印数量
        type: integer
      unprocessed_call_count:
        description: 未处理呼叫数量
        type: integer
    type: object
...
host: 127.0.0.1:8081
info:
  contact: {}
  description: 点餐系统服务端接口文档
  title: ttpos-server-go API
  version: "1.0"
paths:
...
  /cashier/call/abnormal_print/list:
    get:
      consumes:
      - application/json
      description: 异常打印列表
      parameters:
      - description: 页码
        in: query
        name: page_no
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AbnormalPrintList'
              type: object
      security:
      - JwtToken: []
      summary: 异常打印列表
      tags:
      - 收银端.呼叫
  /cashier/call/print:
    delete:
      consumes:
      - application/json
      description: 删除打印
      parameters:
      - description: 删除打印参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.PrinterLogReq'
      produces:
      - application/json
      responses: {}
      security:
      - JwtToken: []
      summary: 删除打印
      tags:
      - 收银端.呼叫
  /cashier/call/processed:
    post:
      consumes:
      - application/json
      description: 处理呼叫
      parameters:
      - description: 处理呼叫参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.ProcessedCallReq'
      produces:
      - application/json
      responses: {}
      security:
      - JwtToken: []
      summary: 处理呼叫
      tags:
      - 收银端.呼叫
  /cashier/call/reprint:
    post:
      consumes:
      - application/json
      description: 重新打印
      parameters:
      - description: 重新打印参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.PrinterLogReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.ReprintResp'
              type: object
      security:
      - JwtToken: []
      summary: 重新打印
      tags:
      - 收银端.呼叫
  /cashier/call/unprocessed:
    get:
      consumes:
      - application/json
      description: 获取未处理消息数量
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.UnprocessedResp'
              type: object
      security:
      - JwtToken: []
      summary: 获取未处理消息数量
      tags:
      - 收银端.呼叫
  /cashier/call/unprocessed/list:
    get:
      consumes:
      - application/json
      description: 未处理呼叫列表
      parameters:
      - description: 页码
        in: query
        name: page_no
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.UnprocessedCallList'
              type: object
      security:
      - JwtToken: []
      summary: 未处理呼叫列表
      tags:
      - 收银端.呼叫
...
  /tablet/call:
    post:
      consumes:
      - application/json
      description: 发起呼叫
      parameters:
      - description: 呼叫请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.CallReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - JwtToken: []
      summary: 发起呼叫
      tags:
      - 平板端.呼叫
...
securityDefinitions:
  JwtToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"